+++ quilt-rs/src/error.rs
@@ -276,6 +276,9 @@ pub enum PackageOpError {
     #[error("Publish error: {0}")]
     Publish(String),
 
+    #[error("Undo error: {0}")]
+    Undo(String),
+
     #[error("General error regarding package: {0}")]
     Package(String),